+++ CHANGELOG.md
@@ -62,6 +62,7 @@ All notable changes to this project will be documented in this file.
 - *(ui)* Improved some dialogs design
 - *(topics)* Added a new solicitation adjustment
 - *(loading)* Adjust loading design for others
+- *(topics)* Added the custom loading for topics cards too
 
 ### 🐛 Bug Fixes